Backend/requirements.txt
Flask
requests
Pillow
flask-cors
beautifulsoup4
